Pull software vertex blending into its own helper
