+++ Sources/Plasma/PubUtilLib/plPipeline/CMakeLists.txt
@@ -11,6 +11,7 @@ set(plPipeline_SOURCES
     plPipelineViewSettings.cpp
     plPlates.cpp
     plRenderTarget.cpp
+    plSoftwareMeshSkinner.cpp
     plStatusLogDrawer.cpp
     plTextFont.cpp
     plTransitionMgr.cpp
@@ -35,6 +36,7 @@ set(plPipeline_HEADERS
     plPipelineViewSettings.h
     plPlates.h
     plRenderTarget.h
+    plSoftwareMeshSkinner.h
     plStatusLogDrawer.h
     plStencil.h
     plTextFont.h
@@ -46,6 +48,7 @@ plasma_library(plPipeline
     SOURCES ${plPipeline_SOURCES} ${plPipeline_HEADERS}
     PRECOMPILED_HEADERS Pch.h
 )
+plasma_target_simd_sources(plPipeline SSE3 plSoftwareMeshSkinner.cpp)
 target_link_libraries(plPipeline
     PUBLIC
         CoreLib
